scripts/requirements.txt
faker
pyvis
psycopg2-binary
